Split coding agent into modules
