+++ src/tiny_llm/agent/__init__.py
@@ -1,0 +1,30 @@
+from .context import append_tool_result, compact_messages
+from .generation import generate_response, initial_messages
+from .loop import AgentEvent, AgentLimits, AgentRun, run_agent
+from .protocol import (
+    AgentError,
+    FinalAction,
+    ToolAction,
+    build_system_prompt,
+    parse_action,
+)
+from .workspace import ToolPolicy, Workspace
+
+
+__all__ = [
+    "AgentError",
+    "AgentEvent",
+    "AgentLimits",
+    "AgentRun",
+    "FinalAction",
+    "ToolAction",
+    "ToolPolicy",
+    "Workspace",
+    "append_tool_result",
+    "build_system_prompt",
+    "compact_messages",
+    "generate_response",
+    "initial_messages",
+    "parse_action",
+    "run_agent",
+]
