+++ src/tiny_llm/agent/context.py
@@ -1,0 +1,15 @@
+from .generation import Message
+
+
+def compact_messages(messages: list[Message], max_chars: int) -> list[Message]:
+    """Week 4, Day 4: retain task anchors and the newest complete tool turns."""
+
+    pass
+
+
+def append_tool_result(
+    messages: list[Message], response: str, result: str
+) -> list[Message]:
+    """Week 4, Day 4: record an action and its observation as one complete turn."""
+
+    pass
